[LibC] Updating types

+++ Userland/Libraries/LibC/include/sys/types.h
@@ -7,6 +7,8 @@
 #ifndef HYPEROS_LIBC_TYPES_H_
 #define HYPEROS_LIBC_TYPES_H_
 
-typedef int pid_t;
+#include <stdint.h>
+
+typedef int64_t pid_t;
 
 #endif // HYPEROS_LIBC_TYPES_H_
